feat(quiz): modularize QuizScreen by adding ChoicesSection, ExplanationSection, QuestionPrompt, and QuizHeader components for better structure and reusability
